Create covariate_script
Script was run in R to create a covariate file containing the covariate values for each individual which was then included in the PRSice2 analysis. The covariate file and the principle component analysis (PCA) were combined and PRSic 2 was run, generating PRS values calculated at various p-value thresholds for the merged sample, with associated plots of the results.