Create SNP_missingness_script
This script was used to identify and list SNPs with different rates of 'missingness' between the case and control samples and consequently any SNPs with a case-control missingness p-value of less than 0.0000I were removed. It was ensured that SNPs were arranged in chromosomal order before EIGESTRAT was used to carry out PCA and produce 100 PCs.