Create pre_qc_steps
First out of the two scripts used for quality control (QC), carried out in PLINK.

QC was carried out to address potential systematic problems in the data, for example, DNA quality issues for
specific individuals, assay issues for specific SNPs, batch effects and sample mishandling.

The quality control protocol, including the relevant thresholds used, were taken from Johan H.
Thygesen and Andries T. Marees et al. (2018). All scripts were run on a server linux using
Bourne-Again SHell (Bash) shell.

The dataset used in the analysis was exported from the GenomeStudio software (Illumina). The
PsychChip raw data was processed into PLINK text format, .ped and fam files, which contain
both genotypic and familial data.

The PED and FAM files were converted into PLINK
binary format, BIM, BED and FAM files, which contain compressed forms of the data and
therefore allowed for a significantly more efficient analysis.


The script also contained commands to clean the data and format it so it would be compatible
with future programs and analyses. The IDs of each participant and their parents were checked
and updated in addition to the phenotype status which was updated to set all subjects as cases.
Minus strand SNPs were flipped to ensure all SNPs were on the positive strand and multi-
matched and SNPs were excluded as their location in the genome could not be confirmed.
Blacklisted SNPs, which had been provided by the Broad institute (MIT) upon their analysis
of the PsychChip data, were excluded. The SNPs on the “Blacklist” were SNPs that were
known to map to multiple regions on the genome or that had produced unreliable genotyping
results. Centimorgan (cM) positions listed in the dataset file were overwritten with zeros to
allow the program EIGENSTRAT to work. New files were then created with the updated
names and positions.