Create qc_steps
SNPs that had a low minor allelic frequency (MAF) were excluded as they are rare and thus
reduce the reliability of their associations with phenotypes. This is particularly significant
given the small size of the dataset. Errors in genotyping are also more likely to arise with low
MAF SNPs. A threshold of 0.01 was used, with SNPs with MAFs below this value excluded.
A threshold was also set for SNP and genome-wide missingness, to reduce both individuals
whose DNA samples were of poor quality and decrease the level of missingness that could lead
to bias in the analysis. Consequently, SNPs undetected in a large majority of subjects and
subjects with large genotype missingness were excluded by a threshold of 0.1 and 0.05
respectively