Create vcf_to_plink
In order for the imputation to work, the PLINK binary files were converted
to VCF format and records were ordered by genomic position. The dataset
was also verified to be compatible with the Genome Reference Consortium Human genome
build 37 (GRCh37). The information (INFO) score of a SNP is a measure of what percentage
of a dataset of the complete sample size, accurately genotyped, the imputed SNP genotypes are
equivalent to. It was calculated to determine how well the SNPs had been imputed and to
consequently remove those that had a low INFO score. A threshold of 0.9 was used to keep
only SNPs of a frequency <1 %. The imputed VCF files were converted to PLINK format using
the --recode command on bash to filter the SNPs.