Create id_prsice_script
Polygenic risk score (PRS) analysis was undertaken using PRSice2, a PRS software which
generates both PRS and visual models of these results. PRSice2 run for each condition
included in the analysis and PRS values were calculated at a various number of p-value
thresholds using the imputed data. The PRS were generated from the risk alleles and weighted
by effect sizes estimated from GWAS results on a base phenotype and the imputed genotype
dataset on a target phenotype. The results were also presented visually through PRS models,
plotted at a range of p-value thresholds.

PRSice2 requires several input files. The genotype data files (BIM, FAM and BED) from the
ID sample, which had undergone quality control procedures and INFO score filtering, were
used as the target dataset. A linkage disequilbrium (LD) reference was also included to improve
estimates of LD during the clumping process, where SNPs that are in LD with each other are
removed. The LD structure in the genotype data from each subject was used. LD clumping was
applied with an R2 threshold of 0.1, where R2 is the percentage of the variation in the
dependent variable that is accounted for by the regression model.