Create ID_prsice_script_anxiety
In PRSice2, the formatting of the columns for chromosome information and the positions of
the SNPs are required to match those of the target file. These differed for the target file and the
anxiety GWAS base file from the Ben Neale Lab GWAS. The Ben Neale results file originally
had a format of “chromosome:position” whereas the target BIM files had a first column of just
chromosome position alone.

This script was made in bash to rectify this and edit the both the base and target file, using the -awk command which searches the files
for these differences and edit them accordingly.